Refactor code as requested
