+++ package.json
@@ -21,12 +21,14 @@
     "dotenv": "^2.0.0",
     "fetch-ponyfill": "^3.0.2",
     "hash.js": "^1.0.3",
+    "json_encode": "^0.1.0",
     "lodash": "^4.17.2",
     "md5-hex": "^1.3.0",
     "moment": "^2.17.1",
     "qs": "^6.3.0",
     "rndm": "^1.2.0",
-    "url": "^0.11.0"
+    "url": "^0.11.0",
+    "urlencode-for-php": "^1.0.2"
   },
   "devDependencies": {
     "ava": "^0.17.0",
