+++ src/index.js
@@ -5,7 +5,7 @@ import Cart from './resources/Cart'
 
 const API_HOST = 'backoffice.2cnnct.com'
 const API_VERSION = 1
-const API_LANG = 'en-gb'
+const API_LANGUAGE = 'en-gb'
 const API_RESELLER_ID = 10000
 
 const resources = {
@@ -19,16 +19,16 @@ export default class Deskbookers {
     https = true,
     host = API_HOST,
     version = API_VERSION,
-    lang = API_LANG,
+    language = API_LANGUAGE,
     resellerId = API_RESELLER_ID
   }) {
     // Options
     this.https = https
     this.host = host
     this.version = version
-    this.lang = lang
+    this.language = language
     this.resellerId = resellerId
-    this.session = null
+    this.session_ = null
 
     // Init resources
     for (let name in resources) {
@@ -37,12 +37,12 @@ export default class Deskbookers {
   }
 
   hasSession () {
-    return !!(this.session && this.session.privateKey)
+    return !!(this.session_ && this.session_.privateKey)
   }
 
-  getSession () {
+  get session () {
     return this.hasSession()
-      ? { ...this.session }
+      ? { ...this.session_ }
       : null
   }
 
@@ -65,34 +65,34 @@ export default class Deskbookers {
     }
   }
 
-  setSession (session) {
+  set session (session) {
     if (
       session &&
       session.privateKey &&
       session.publicKey &&
       session.user
     ) {
-      this.session = session
+      this.session_ = session
     } else {
-      this.session = null
+      this.session_ = null
     }
     return this
   }
 
   async logout () {
-    this.setSession()
+    this.session = null
     return this
   }
 
   async login (email, password, ...args) {
     const session = await this.users.login(email, password, ...args)
-    this.setSession(session)
+    this.session = session
     return session.user
   }
 
   async signup (params, ...args) {
     const session = await this.users.signup(params, ...args)
-    this.setSession(session)
+    this.session = session
     return session.user
   }
 }
