+++ src/resources/Resource.js
@@ -1,5 +1,5 @@
-import fetch from '../lib/fetch'
-import { signer, formatArgs } from '../lib/request-utils'
+import fetch from '../utils/fetch'
+import { signer, formatArgs } from '../utils/requests'
 
 export default class Resource {
   constructor (api) {
@@ -20,19 +20,20 @@ export default class Resource {
 
     const args = {
       __resellerID: this.api.resellerId,
-      __i18n: this.api.lang,
+      __i18n: this.api.language,
       __fields: fields,
       ...params
     }
     const queryStr = formatArgs(args, options.method === 'post')
+    const pathFixed = path.replace(/^\/+|\/+$/, '')
 
     let url
     if (options.method === 'post') {
       options.headers['Content-Type'] = 'application/x-www-form-urlencoded;charset=UTF-8'
       options.body = queryStr
-      url = `${this.apiUrl}/${path.replace(/^\/+|\/+$/, '')}`
+      url = `${this.apiUrl}/${pathFixed}`
     } else {
-      url = `${this.apiUrl}/${path.replace(/^\/+|\/+$/, '')}?${queryStr}`
+      url = `${this.apiUrl}/${pathFixed}?${queryStr}`
     }
 
     const {
@@ -51,49 +52,35 @@ export default class Resource {
 
   async parseResponse (response) {
     // Retrieve json data
-    let data
-    try {
-      data = await response.text()
-      data = JSON.parse(data)
-    } catch (e) {
-      console.log('Invalid JSON result received:\n', data)
-      console.error(e)
-      throw e
-    }
+    const data = await response.json()
 
     // Parse response data
-    try {
-      if (data && typeof data === 'object') {
-        if (data.error) {
-          const msg = data.errorMessage || 'An error occurred'
-          const extra = data.errorTrace && data.errorTrace.length > 0
-            ? ` (${data.errorTrace[0].file}:${data.errorTrace[0].line})`
-            : null
-          throw new Error(`${data.errorCode || 500}: ${msg}${extra}`)
-        } else if (data.errors && data.errors.length > 0) {
-          for (let i = 0; i < data.errors.length; ++i) {
-            const msg = data.errors[i].title || data.errors[i].detail
-            console.error(data.errors[i])
-            throw new Error(`${data.errors[i].code || 500}: ${msg}`)
-          }
-        } else if ('result' in data) return data.result
-      }
-
-      // Error
-      console.error('Invalid response format:', data)
-      throw new Error('500: Invalid response received')
-    } catch (e) {
-      console.error(e)
-      throw e
+    if (data && typeof data === 'object') {
+      if (data.error) {
+        const msg = data.errorMessage || 'An error occurred'
+        const extra = data.errorTrace && data.errorTrace.length > 0
+          ? ` (${data.errorTrace[0].file}:${data.errorTrace[0].line})`
+          : null
+        throw new Error(`${data.errorCode || 500}: ${msg}${extra}`)
+      } else if (data.errors && data.errors.length > 0) {
+        for (let i = 0; i < data.errors.length; ++i) {
+          const msg = data.errors[i].title || data.errors[i].detail
+          console.error(data.errors[i])
+          throw new Error(`${data.errors[i].code || 500}: ${msg}`)
+        }
+      } else if ('result' in data) return data.result
     }
+
+    // Error
+    throw new Error('500: Invalid response received')
   }
 
   async prepareRequest (url, options, args) {
+    // Sign request?
     if (this.api.hasSession()) {
-      return signer(this.api.getSession(), url, options, args)
+      return signer(this.api.session, url, options, args)
+    } else {
+      return { url, options }
     }
-
-    // Or don't sign
-    return { url, options }
   }
 }
